Auth reworks I made
First I removed the utils folder. Second I am adding in tests for api calls in login folder. Added in constants folder, so I can save the constant action types to clean up the code, make it more readable. Debating on removing to clean up the code. Going to rework the reducer. Now trying to configure the api call to check out what is given in the response.